Split Tensormesh chat test helpers
